Somando dois numeros (inteiros)

+++ ex004.py
@@ -1,4 +1,4 @@
-n1 = float(input("Digite um numero: "))
-n2 = float(input("Digite outro numero: "))
+n1 = int(input("Digite um numero: "))
+n2 = int(input("Digite outro numero: "))
 s = n1 + n2
 print("A soma entre {} e {} é {}".format(n1, n2, s))
